+++ Sources/PackageConfig/Loader.swift
@@ -2,9 +2,8 @@
 import Foundation
 
 enum Loader {
-	static func load<T: Configuration>() -> T? {
-		let path = NSTemporaryDirectory()
-		let packageConfigJSON = path + "package-config.json"
+	static func load<T: PackageConfig>() -> T? {
+		let packageConfigJSON = NSTemporaryDirectory() + "package-config.json"
 
 		guard let data = FileManager.default.contents(atPath: packageConfigJSON) else {
 			debugLog("Could not find a file at \(packageConfigJSON) - so returning an empty object")
