+++ Sources/PackageConfig/Writer.swift
@@ -3,16 +3,15 @@ import Foundation
 
 enum Writer {
 
-	static func write<T: Codable>(configuration: T) { //}, adapter: TypePreservingCodingAdapter) {
-		let fileManager = FileManager.default
-		let path = NSTemporaryDirectory()
-		let packageConfigJSON = path + "package-config.json"
+	static func write<T: Codable>(configuration: T) {
+		let packageConfigJSON = NSTemporaryDirectory() + "package-config.json"
 		let encoder = JSONEncoder()
-//		encoder.userInfo[.typePreservingAdapter] = adapter
 		let data = try! encoder.encode(configuration)
 
-		if !fileManager.createFile(atPath: packageConfigJSON, contents: data, attributes: nil) {
+		if !FileManager.default.createFile(atPath: packageConfigJSON, contents: data, attributes: nil) {
 			print("PackageConfig: Could not create a temporary file for the PackageConfig: \(packageConfigJSON)")
 		}
+
+		debugLog("written to path: \(packageConfigJSON)")
 	}
 }
